+++ src/main/java/com/ocho/what2do/store/service/KakaoApiServiceImpl.java
@@ -1,7 +1,7 @@
-package com.ocho.what2do.common.daum.service;
+package com.ocho.what2do.store.service;
 
-import com.ocho.what2do.common.daum.entity.ApiStore;
-import com.ocho.what2do.common.daum.repository.ApiStoreRepository;
+import com.ocho.what2do.store.entity.ApiStore;
+import com.ocho.what2do.store.repository.ApiStoreRepository;
 import com.ocho.what2do.common.exception.CustomException;
 import com.ocho.what2do.common.message.CustomErrorCode;
 import com.ocho.what2do.store.dto.StoreListResponseDto;
@@ -11,6 +11,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -24,44 +25,9 @@ import java.util.List;
 @Slf4j(topic = "DAUM API")
 @Service
 @RequiredArgsConstructor
-public class DaumApiServiceImpl implements DaumApiService {
+public class KakaoApiServiceImpl implements KakaoApiService {
 
-    private final RestTemplate restTemplate;
     private final ApiStoreRepository apiStoreRepository;
-
-    @Value("${kakao.Authorization}")
-    private String Authorization;
-
-    @Override
-//    @TimeTrace
-    public StoreListResponseDto searchItems(String query, String page) {
-        // 요청 URL 만들기
-        URI uri = UriComponentsBuilder
-                .fromUriString("https://dapi.kakao.com")
-                .path("/v2/local/search/keyword.json")
-                .queryParam("query", query) // 검색어
-                .queryParam("page", page) // 페이징
-                .queryParam("size", "15") // 한 페이지에 출력 할 값의 갯수 (고정 값 설정)
-                .queryParam("category_group_code", "FD6") // FD6 음식점 코드 (음식점만 검색 결과로 나오게끔)
-                .encode()
-                .build()
-                .toUri();
-        log.info("uri = " + uri);
-
-        RequestEntity<Void> requestEntity = RequestEntity
-                .get(uri)
-                .header("Authorization", Authorization)
-                .build();
-
-        ResponseEntity<String> responseEntity = restTemplate.exchange(requestEntity, String.class);
-
-        log.info("DAUM API Status Code : " + responseEntity.getStatusCode());
-
-        if (Integer.parseInt(page) > 3) {
-            throw new CustomException(CustomErrorCode.DATA_NOT_FOUND);
-        }
-        return fromJSONtoItems(responseEntity.getBody());
-    }
 
     @Override
     public StoreListResponseDto fromJSONtoItems(String responseEntity) {
@@ -93,10 +59,14 @@ public class DaumApiServiceImpl implements DaumApiService {
                     .latitude(storeResponseDto.getLatitude())
                     .longitude(storeResponseDto.getLongitude())
                     .build();
-            if (!apiStoreRepository.existsApiStoreByStoreKey(store.getStoreKey())) {
-                apiStoreRepository.save(store);
-            }
             storeResponseDtoList.add(storeResponseDto);
+            try {
+                if (!apiStoreRepository.existsApiStoreByStoreKey(store.getStoreKey())) {
+                    apiStoreRepository.save(store);
+                }
+            } catch (DataIntegrityViolationException e) {
+                continue;
+            }
         }
         return new StoreListResponseDto(totalCnt, pageCnt, keyWord, region, storeResponseDtoList);
     }
